skip symlinks if they already exists, fix subject list default

+++ workflow/scripts/symlink_dicoms_by_field_strength.py
@@ -4,13 +4,13 @@ import pydicom
 from pathlib import Path
 from bidscoin import lsdirs
 
-def symlink_dicoms_by_field_strength(source_dicoms_folder: str, output_folder: str, subject_list=None):
+def symlink_dicoms_by_field_strength(source_dicoms_folder: str, output_folder: str, subject_list="*"):
     # Convert source folder to Path object
     rawfolder = Path(source_dicoms_folder).resolve()
     
     # List subjects folders, assuming they are the next directories in the hierarchy
     subjects = []
-    if subject_list is not None:
+    if subject_list != "*" :
         for sub in subject_list:
             subject_folder = lsdirs(rawfolder, sub)[0]
             subjects.append(subject_folder)
@@ -44,7 +44,8 @@ def symlink_dicoms_by_field_strength(source_dicoms_folder: str, output_folder: s
             
             #symlink the contents of the session folder to the new location
             for item in os.listdir(session):
-                os.symlink(os.path.join(session, item), os.path.join(new_folder, item))
+                if not os.path.exists(os.path.join(new_folder, item)):
+                    os.symlink(os.path.join(session, item), os.path.join(new_folder, item))
 
             #remove dicoms with NoAV in the name
             for dicom in list(new_folder.rglob('*NoAV*.dcm', recurse_symlinks=True)):
@@ -55,6 +56,6 @@ if __name__ == '__main__':
         description="Symlink DICOM files to new folders based on their magnetic field strength")
     parser.add_argument('source_dicoms_folder', type=str, help="Path to the source DICOM folder")
     parser.add_argument('output_folder', type=str, help="Path to the output folder where 3T and 7T folders will be created and DICOMS copied")
-    parser.add_argument('subject_list', nargs="*", type=str, default=None, help="Space separated list of subject folder names. Default is all subject folders in the source DICOM folder.")
+    parser.add_argument('subject_list', nargs="*", type=str, default="*", help="Space separated list of subject folder names. Default is all subject folders in the source DICOM folder.")
     args = parser.parse_args()
-    copy_dicoms_by_field_strength(args.source_dicoms_folder, args.output_folder, args.subject_list)
+    symlink_dicoms_by_field_strength(args.source_dicoms_folder, args.output_folder, args.subject_list)
